requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-js-asset==2.0.0
django-mptt==0.13.4
Pillow==9.2.0
psycopg2-binary==2.9.3
six==1.16.0
sqlparse==0.4.2
validate-email==1.3
